Replace renderStyling with PostCSS parser
renderStyling was mostly concerned with transforming a JS object into a CSS string. Because it used a simple approach, it wasn't able to handle special cases. By using PostCSS we can completely ignore the aspect of how to parse a JS object into a valid CSS string plus have the option to use a number of PostCSS plugins later on.

So yes, this introduces an additional dependency, but it should be worth it, as it further standardizes CSS in JS.