Example of how to use the Postcss nesting plugin
This example shows how additional Postcss plugins can be used by using the Postcss loader and specifying plugins it should use.

This is likely an unnecessary step, because Postcss will run twice with this setup. Instead, it should be possible to specify Postcss plugins that Styling should use directly when it does the transformation.